# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CPay API
  version: v1
servers:
  - url: http://localhost:8080
paths:
...
  /v1/payment_links/{id}/sessions:
    post:
      summary: Create checkout session from payment link
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key